requirements.txt
opencv-python
pillow
pyzbar
pypylon
numpy
pysqlite3
pyserial